Update view module, docs
