+++ doc/configuration.rst
@@ -1,0 +1,12 @@
+configuration
+==============
+
+multilog is configured using a file 'config.yml', that has to be placed in the main directory.
+
+For a detailed description see https://github.com/nemocrys/multilog#configuration
+
+template config file
+--------------------
+
+.. literalinclude:: ../config_template.yml
+
